refs #22031 improved error checking in FDA
