+++ scripts/Muon/thread_model.py
@@ -1,6 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 
 from PyQt4.QtCore import QThread
+import mantid.simpleapi as mantid
 
 
 class ThreadModel(QThread):
@@ -9,18 +10,35 @@ class ThreadModel(QThread):
     A wrapper to allow threading with
     the MaxEnt models.
     """
-    def __init__(self,model):
+    def __init__(self,model,run):
         QThread.__init__(self)
         self.model=model
+        self.run = run
 
     def __del__(self):
         self.wait()
+ 
+    def digit(self,x):
+        return int(filter(str.isdigit,x) or 0)
+
+    def hasDataChanged(self):
+        ws = mantid.AnalysisDataService.retrieve("MuonAnalysis")
+        if mantid.AnalysisDataService.doesExist("MuonAnalysis_1"):
+            ws = mantid.AnalysisDataService.retrieve("MuonAnalysis_1")
+
+        current = ws.getInstrument().getName()+str(ws.getRunNumber()).zfill(8)
+        if self.run != current:
+            mantid.logger.error("Active workspace has changed. Restart this interface")
+            return True
+        return False
 
     def run(self):
-        try:
+       if self.hasDataChanged():
+           return
+       try:
             self.model.execute()
             self.model.output()
-        except KeyboardInterrupt:
+       except KeyboardInterrupt:
             pass
 
     def cancel(self):
